Move out profiler logic to mixin
