+++ corehq/apps/reports/standard/cases/basic.py
@@ -14,6 +14,7 @@ from corehq.apps.reports.generic import ElasticProjectInspectionReport
 from corehq.apps.reports.standard import (
     ProjectReport,
     ProjectReportParametersMixin,
+    ESQueryProfilerMixin,
 )
 from corehq.apps.reports.standard.cases.filters import CaseSearchFilter
 from corehq.apps.reports.standard.cases.utils import (
@@ -26,10 +27,9 @@ from corehq.elastic import ESError
 from corehq.util.es.elasticsearch import TransportError
 
 from .data_sources import CaseDisplayES
-from corehq.apps.case_search.utils import ESQueryProfiler
 
 
-class CaseListMixin(ElasticProjectInspectionReport, ProjectReportParametersMixin):
+class CaseListMixin(ESQueryProfilerMixin, ElasticProjectInspectionReport, ProjectReportParametersMixin):
     fields = [
         'corehq.apps.reports.filters.case_list.CaseListFilter',
         'corehq.apps.reports.filters.select.CaseTypeFilter',
@@ -40,29 +40,7 @@ class CaseListMixin(ElasticProjectInspectionReport, ProjectReportParametersMixin
     case_filter = {}
     ajax_pagination = True
     asynchronous = True
-    es_search_class = case_es.CaseES
-    profiler_enabled = False
-    profiler_class = ESQueryProfiler
-    profiler_name = 'Case List'
-    _profiler = None
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        if self.profiler_enabled:
-            self._profiler = self.profiler_class(
-                name=self.profiler_name,
-                search_class=self.es_search_class,
-            )
-
-    @property
-    def profiler(self):
-        return self._profiler
-
-    @property
-    def search_class(self):
-        if not self.profiler_enabled:
-            return self.es_search_class
-        return self.profiler.get_search_class(slug=self.__class__.__name__)
+    search_class = case_es.CaseES
 
     def _base_query(self):
         return (
