+++ corehq/apps/reports/standard/cases/case_list_explorer.py
@@ -57,7 +57,7 @@ class XpathCaseSearchFilterMixin(object):
 class CaseListExplorer(CaseListReport, XpathCaseSearchFilterMixin):
     name = _('Case List Explorer')
     slug = 'case_list_explorer'
-    es_search_class = CaseSearchES
+    search_class = CaseSearchES
     description = _("Use Case List Explorer to run deep searches on your cases by case properties.  ")
     documentation_link = DOCS_LINK_CASE_LIST_EXPLORER
     use_bootstrap5 = False
